revert unneeded js changes
