+++ src/dymaptic.GeoBlazor.Core/Scripts/basemap.gb.ts
@@ -208,6 +208,57 @@ export default class BasemapGenerated implements IPropertyWrapper {
 }
 
 
+export async function buildJsBasemapGenerated(dotNetObject: any, layerId: string | null, viewId: string | null): Promise<any> {
+    if (!hasValue(dotNetObject)) {
+        return null;
+    }
+
+    let properties: any = {};
+    if (hasValue(dotNetObject.baseLayers) && dotNetObject.baseLayers.length > 0) {
+        let { buildJsLayer } = await import('./layer');
+        properties.baseLayers = await Promise.all(dotNetObject.baseLayers.map(async i => await buildJsLayer(i, layerId, viewId))) as any;
+    }
+    if (hasValue(dotNetObject.portalItem)) {
+        let { buildJsPortalItem } = await import('./portalItem');
+        properties.portalItem = await buildJsPortalItem(dotNetObject.portalItem, layerId, viewId) as any;
+    }
+    if (hasValue(dotNetObject.referenceLayers) && dotNetObject.referenceLayers.length > 0) {
+        let { buildJsLayer } = await import('./layer');
+        properties.referenceLayers = await Promise.all(dotNetObject.referenceLayers.map(async i => await buildJsLayer(i, layerId, viewId))) as any;
+    }
+    if (hasValue(dotNetObject.spatialReference)) {
+        let { buildJsSpatialReference } = await import('./spatialReference');
+        properties.spatialReference = buildJsSpatialReference(dotNetObject.spatialReference) as any;
+    }
+    if (hasValue(dotNetObject.style)) {
+        let { buildJsBasemapStyle } = await import('./basemapStyle');
+        properties.style = await buildJsBasemapStyle(dotNetObject.style) as any;
+    }
+
+    if (hasValue(dotNetObject.basemapId)) {
+        properties.id = dotNetObject.basemapId;
+    }
+    if (hasValue(dotNetObject.thumbnailUrl)) {
+        properties.thumbnailUrl = dotNetObject.thumbnailUrl;
+    }
+    if (hasValue(dotNetObject.title)) {
+        properties.title = dotNetObject.title;
+    }
+    let jsBasemap = new Basemap(properties);
+
+    let { default: BasemapWrapper } = await import('./basemap');
+    let basemapWrapper = new BasemapWrapper(jsBasemap);
+    basemapWrapper.geoBlazorId = dotNetObject.id;
+    basemapWrapper.viewId = viewId;
+    basemapWrapper.layerId = layerId;
+    
+    jsObjectRefs[dotNetObject.id] = basemapWrapper;
+    arcGisObjectRefs[dotNetObject.id] = jsBasemap;
+    
+    return jsBasemap;
+}
+
+
 export async function buildDotNetBasemapGenerated(jsObject: any): Promise<any> {
     if (!hasValue(jsObject)) {
         return null;
